tests for properties
