+++ Assets/Tests/Runtime/Serialization/BitPackingProperties.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a96aaa0a5f3acb4a83b659c28316658
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
