+++ Assets/Tests/Runtime/Serialization/BitPackingReusingWriterTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d600aaf09129e234c9ccd9b98223cdc9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
