+++ Assets/Tests/Runtime/Serialization/BitPackingTests.cs
@@ -4,95 +4,6 @@ using NUnit.Framework;
 
 namespace Mirage.Tests.Runtime.Serialization
 {
-    public class BitPackingReusingWriterTests
-    {
-        private NetworkWriter writer;
-        private NetworkReader reader;
-
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            writer = new NetworkWriter(1300);
-            reader = new NetworkReader();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            writer.Reset();
-            reader.Dispose();
-        }
-
-
-        [Test]
-        public void WriteUShortAfterReset()
-        {
-            ushort value1 = 0b0101;
-            ushort value2 = 0x1000;
-
-            // write first value
-            writer.WriteUInt16(value1);
-
-            reader.Reset(writer.ToArray());
-            ushort out1 = reader.ReadUInt16();
-            Assert.That(out1, Is.EqualTo(value1));
-
-            // reset and write 2nd value
-            writer.Reset();
-
-            writer.WriteUInt16(value2);
-
-            reader.Reset(writer.ToArray());
-            ushort out2 = reader.ReadUInt16();
-            Assert.That(out2, Is.EqualTo(value2), "Value 2 was incorrect");
-        }
-
-        [Test]
-        [TestCase(0b0101ul, 0x1000ul)]
-        [TestCase(0xffff_0000_ffff_fffful, 0x0000_ffff_1111_0000ul)]
-        public void WriteULongAfterReset(ulong value1, ulong value2)
-        {
-            // write first value
-            writer.WriteUInt64(value1);
-
-            reader.Reset(writer.ToArray());
-            ulong out1 = reader.ReadUInt64();
-            Assert.That(out1, Is.EqualTo(value1));
-
-            // reset and write 2nd value
-            writer.Reset();
-
-            writer.WriteUInt64(value2);
-
-            reader.Reset(writer.ToArray());
-            ulong out2 = reader.ReadUInt64();
-            Assert.That(out2, Is.EqualTo(value2), "Value 2 was incorrect");
-        }
-
-        [Test]
-        [TestCase(0b0101ul, 0x1000ul)]
-        [TestCase(0xffff_0000_ffff_fffful, 0x0000_ffff_1111_0000ul)]
-        public void WriteULongWriteBitsAfterReset(ulong value1, ulong value2)
-        {
-            // write first value
-            writer.Write(value1, 64);
-
-            reader.Reset(writer.ToArray());
-            ulong out1 = reader.Read(64);
-            Assert.That(out1, Is.EqualTo(value1));
-
-            // reset and write 2nd value
-            writer.Reset();
-
-            writer.Write(value2, 64);
-
-            reader.Reset(writer.ToArray());
-            ulong out2 = reader.Read(64);
-            Assert.That(out2, Is.EqualTo(value2), "Value 2 was incorrect");
-        }
-    }
-
-
     public class BitPackingTests
     {
         private NetworkWriter writer;
@@ -208,54 +119,6 @@ namespace Mirage.Tests.Runtime.Serialization
             _ = reader.Read(writerBits);
             ulong result = reader.Read(valueBits);
             Assert.That(result, Is.EqualTo(maskedValue));
-        }
-    }
-    public class BitMaskHelperTests
-    {
-        /// <summary>
-        /// slow way of creating correct mask
-        /// </summary>
-        static ulong slowMask(int bits)
-        {
-            ulong mask = 0;
-            for (int i = 0; i < bits; i++)
-            {
-                mask |= 1ul << i;
-            }
-
-            return mask;
-        }
-
-        [Test]
-        [Description("manually checking edge cases to be sure")]
-        public void MaskValueIsCorrect0()
-        {
-            ulong mask = BitMask.Mask(0);
-            Assert.That(mask, Is.EqualTo(0x0));
-        }
-
-        [Test]
-        [Description("manually checking edge cases to be sure")]
-        public void MaskValueIsCorrect63()
-        {
-            ulong mask = BitMask.Mask(63);
-            Assert.That(mask, Is.EqualTo(0x7FFF_FFFF_FFFF_FFFF));
-        }
-
-        [Test]
-        [Description("manually checking edge cases to be sure")]
-        public void MaskValueIsCorrect64()
-        {
-            ulong mask = BitMask.Mask(64);
-            Assert.That(mask, Is.EqualTo(0xFFFF_FFFF_FFFF_FFFF));
-        }
-
-        [Test]
-        public void MaskValueIsCorrect([Range(0, 64)] int bits)
-        {
-            ulong mask = BitMask.Mask(bits);
-            ulong expected = slowMask(bits);
-            Assert.That(mask, Is.EqualTo(expected), $"    mask:{mask:X}\nexpected:{expected:X}");
         }
     }
